refactor(1013): extract GatewayTestExporter (SC-001 position 37) (#1013)
Extract GatewayTestExporter class (~245 LOC, 14 static methods) from
MistHelper.py to src/export/gateway_test_exporter.py using the Cat B
SC-001 facade pattern:

- Lazy `mh = importlib.import_module("MistHelper")` inside every helper
  that touches a live global (apisession, ConfigUtils,
  GatewayExportUtils, ValidationUtils, ConnectionPoolExecutor,
  RateLimitingUtils, DataProcessingUtils, DataExporter, PROGRESS_EMITTER,
  ProgressContext, FAST_MODE_* / FastModeBackoffMultiplier /
  FastModeSequentialMaxRetries, _api_usage_cache).
- Re-export alias added in MistHelper.py preserving
  `MistHelper.GatewayTestExporter` callers.
- `test_results_by_site` retained as thin delegator to
  `GatewayTestResultsService`.
- Removed newly-unused `as_completed` top-level import from
  MistHelper.py (last consumer moved out).
- pyproject coverage omit updated; `src.export.*` already in Wave 2
  mypy override list.

Local gate:
- black --check + ruff check clean on both files
- py_compile clean
- python MistHelper.py --test baseline: 66 pass / 0 fail / 131 skip

Refs #1013.
