+++ MistHelper.py
@@ -50,8 +50,7 @@ import traceback  # Import traceback for detailed exception context in error log
 from collections.abc import Callable  # Import Callable type hint for callback functions passed to API methods
 from concurrent.futures import (
     ThreadPoolExecutor,
-    as_completed,
-)  # Thread-pool primitives (FIRST_COMPLETED/wait migrated to ConnectionPoolExecutor per 1012 SC-003)
+)  # Thread-pool primitive (as_completed migrated per 1013 P37; FIRST_COMPLETED/wait per 1012 SC-003)
 from dataclasses import dataclass, field  # Import dataclass decorators for configuration objects and entity classes
 from datetime import datetime  # Import datetime for timestamping logs and events
 from typing import TYPE_CHECKING, Any, Literal  # Import type hints for static analysis without runtime overhead
@@ -142,6 +141,9 @@ from src.export.const_definitions_exporter import (  # pylint: disable=unused-im
 )
 from src.export.device_events_52w_exporter import (  # pylint: disable=unused-import
     DeviceEvents52wExporter,  # noqa: F401  # Re-export preserved after OrgAlarmEventExporter extraction (1013 SC-001 position 18)
+)
+from src.export.gateway_test_exporter import (  # pylint: disable=unused-import
+    GatewayTestExporter,  # noqa: F401  # Cat B (1013 SC-001 position 37) -- re-export for MistHelper.GatewayTestExporter callers
 )
 from src.export.license_export_utils import (  # pylint: disable=unused-import
     LicenseExportUtils,  # noqa: F401  # Cat B (1013 SC-001 position 24) -- re-export for MistHelper.LicenseExportUtils callers
@@ -10978,254 +10980,7 @@ class InsightMetricsUtils:  # Insight-metrics helpers.
 # src/ui/interactive_display_utils.py (issue #1013 SC-001 position 10)
 
 
-# ============================================================================
-# GATEWAY EXPORT UTILITIES CLASS
-# ============================================================================
-class GatewayTestExporter:  # Gateway synthetic test exporter.
-    """
-    Gateway Synthetic Test Exports
-
-    Handles synthetic test result exports and site-level test aggregation for gateways.
-    Extracted from GatewayExportUtils.
-    """
-
-    @staticmethod
-    def synthetic_tests(fast=False):
-        """Collect + export synthetic test stats for all gateways (optional fast/concurrent path)."""
-        logging.debug("[DEBUG] GatewayTestExporter.synthetic_tests invoked with fast=%s", fast)  # Entry trace.
-        logging.info("[INFO] Collecting synthetic test stats for all gateways in the org...")  # Log start.
-        if fast:  # Fast mode.
-            logging.info(" Fast mode enabled: Using cached data and concurrent processing (synthetic tests)")
-        emitter = PROGRESS_EMITTER  # Progress emitter.
-        if emitter:  # Emitter present.
-            emitter.emit_progress_start("16", "synthetic_tests", 1)  # Signal progress start.
-        op_start = time.time()  # Start the timer.
-        org_id = ConfigUtils.get_cached_or_prompted_org_id()  # Resolve the org.
-        gateway_devices = GatewayExportUtils._get_devices_with_sites(org_id, fast=fast)  # List gateways.
-        if not gateway_devices:  # No gateways.
-            logging.warning("[WARN] No gateway devices found. Exiting synthetic tests export.")  # Warn.
-            return  # Abort.
-        all_stats: list = []  # Accumulate stats.
-        if fast:  # Concurrent path.
-            GatewayTestExporter._run_synthetic_fast_path(gateway_devices, all_stats)  # Fast pool.
-        else:
-            GatewayTestExporter._run_synthetic_sequential_path(gateway_devices, all_stats)  # Sequential.
-        GatewayTestExporter._export_synthetic_results(all_stats, gateway_devices)  # Write CSV + log.
-        GatewayTestExporter._emit_synthetic_complete(emitter, op_start, gateway_devices, all_stats)  # Done.
-
-    @staticmethod
-    def _emit_synthetic_complete(emitter, op_start, gateway_devices, all_stats):
-        """Emit final progress-complete signal if an emitter is configured."""
-        if not emitter:  # No emitter — nothing to emit.
-            return  # Skip silently.
-        emitter.emit_progress_complete(  # Signal progress complete.
-            ProgressContext("16", "synthetic_tests", len(gateway_devices)),
-            len(all_stats),
-            False,
-            time.time() - op_start,
-        )
-
-    @staticmethod
-    def _resolve_retry_defaults(max_retries, retry_delay):  # type: ignore[no-untyped-def]
-        """Apply FAST_MODE defaults for unset retry budget / delay (returns the tuple)."""
-        if max_retries is None:  # Default max retries.
-            max_retries = FAST_MODE_MAX_RETRIES  # Fast-mode default.
-        if retry_delay is None:  # Default retry delay.
-            retry_delay = FAST_MODE_RETRY_DELAY  # Fast-mode default.
-        return max_retries, retry_delay  # Tuple back to caller
-
-    @staticmethod
-    def fetch_synthetic_test_stats_with_retry(
-        device_info, max_retries=None, retry_delay=None, connection_semaphore=None
-    ):
-        """Fetch synthetic test stats for one gateway with retry + optional connection pool gating."""
-        max_retries, retry_delay = GatewayTestExporter._resolve_retry_defaults(
-            max_retries, retry_delay
-        )  # Defaults via helper
-        site_id, device_id, _, _ = device_info  # Unpack for backoff/error logs only (names unused here).
-        for attempt in range(max_retries + 1):  # Bounded retry loop.
-            stats = GatewayTestExporter._try_synthetic_fetch_attempt(
-                device_info, attempt, connection_semaphore
-            )  # One attempt.
-            if stats is not None:  # Success.
-                return stats  # Return tagged stats.
-            if attempt >= max_retries:  # Out of retries.
-                logging.error("! Final attempt failed for device %s at site %s", device_id, site_id)  # Final failure.
-                return None  # Give up.
-            backoff_delay = retry_delay * (FastModeBackoffMultiplier.VALUE**attempt)  # Exponential backoff.
-            logging.info(  # Log the retry.
-                "! Fast retry in %.1fs (attempt %s/%s)", backoff_delay, attempt + 2, max_retries + 1
-            )
-            time.sleep(backoff_delay)  # Wait before retry.
-        return None  # Defensive.
-
-    @staticmethod
-    def _try_synthetic_fetch_attempt(device_info, attempt, connection_semaphore):
-        """Single attempt: validate inputs, call API, tag stats, log success. Return stats or None."""
-        site_id, device_id, _, _ = device_info  # Unpack for the call + logging (names unused directly here).
-        try:
-            ValidationUtils.validate_site_id(site_id, "synthetic_tests")  # Validate the site id.
-            ValidationUtils.validate_device_id(device_id, "synthetic_tests")  # Validate the device id.
-            stats = GatewayTestExporter._call_synthetic_endpoint(site_id, device_id, connection_semaphore)  # Call API.
-            GatewayTestExporter._tag_synthetic_stats(stats, device_info, attempt)  # Tag + log success.
-            return stats  # Return tagged stats.
-        except Exception as exception:  # Fetch failed this attempt.
-            logging.warning(  # Warn and let caller handle backoff/retry.
-                "! Attempt %s failed for device %s at site %s: %s",
-                attempt + 1,
-                device_id,
-                site_id,
-                exception,
-            )
-            return None  # Signal failure to caller.
-
-    @staticmethod
-    def _tag_synthetic_stats(stats, device_info, attempt):
-        """Mutate ``stats`` with site/device tags and log first-try vs retry success."""
-        site_id, device_id, device_name, site_name = device_info  # Unpack the fields.
-        stats["site_id"] = site_id  # Tag the site.
-        stats["site_name"] = site_name  # Tag the site name.
-        stats["device_id"] = device_id  # Tag the device.
-        stats["device_name"] = device_name  # Tag the device name.
-        if attempt > 0:  # After a retry.
-            logging.info("! Retry %s successful for device %s at site %s", attempt, device_name, site_name)
-        else:
-            logging.info("! Collected synthetic test stats for device %s at site %s", device_name, site_name)
-
-    @staticmethod
-    def _call_synthetic_endpoint(site_id, device_id, connection_semaphore):
-        """Call ``getSiteDeviceSyntheticTest`` with optional semaphore-gated concurrency."""
-        if connection_semaphore:  # Pool present.
-            with connection_semaphore:  # Acquire a slot.
-                return mistapi.api.v1.sites.devices.getSiteDeviceSyntheticTest(apisession, site_id, device_id).data
-        return mistapi.api.v1.sites.devices.getSiteDeviceSyntheticTest(
-            apisession, site_id, device_id
-        ).data  # Unsemaphored call.
-
-    @staticmethod
-    def _run_synthetic_fast_path(gateway_devices, all_stats):
-        """Concurrent pool execution with retry on failures + summary instrumentation."""
-        start_time = time.time()  # Start the timer.
-
-        def fetch_device_stats(device_info, connection_semaphore):  # Pool worker.
-            """Worker function that fetches synthetic test stats for a single device."""
-            return GatewayTestExporter.fetch_synthetic_test_stats_with_retry(
-                device_info, connection_semaphore=connection_semaphore
-            )
-
-        successful_results, failed_devices = ConnectionPoolExecutor.execute(  # Pool run.
-            work_items=gateway_devices,
-            worker_function=fetch_device_stats,
-            batch_description="devices",
-            retry_function=GatewayTestExporter._retry_failed_synthetic_devices,
-        )
-        duration = time.time() - start_time  # Compute the duration.
-        all_stats.extend(successful_results)  # Collect the results.
-        logging.info(  # Log the totals.
-            " FAST MODE SUMMARY (synthetic tests): ok=%s fail=%s total=%s elapsed=%.2fs",
-            len(successful_results),
-            len(failed_devices),
-            len(gateway_devices),
-            duration,
-        )
-
-    @staticmethod
-    def _retry_failed_synthetic_devices(failed_devices, connection_semaphore):
-        """Retry failed devices through a small dedicated pool. Return (results, still_failed)."""
-        retry_threads = min(  # Size the retry pool.
-            FAST_MODE_RETRY_THREADS,
-            len(failed_devices),
-            max(1, FAST_MODE_MAX_CONCURRENT_CONNECTIONS - 2),
-        )
-        if retry_threads <= 0:  # No threads available.
-            logging.warning(" FAST MODE: No available threads for retry; skipping retries")
-            return [], failed_devices  # Return original failures.
-        retry_results: list = []  # Collect retry results.
-        still_failed: list = []  # Track still-failed devices.
-        with ThreadPoolExecutor(max_workers=retry_threads) as executor:  # Run the retry pool.
-            retry_futures = GatewayTestExporter._submit_synthetic_retries(
-                executor, failed_devices, connection_semaphore
-            )  # Build future map.
-            for future in tqdm(  # type: ignore[no-untyped-call]
-                as_completed(retry_futures),
-                total=len(retry_futures),
-                desc="Retrying Failed",
-                unit="device",
-            ):
-                GatewayTestExporter._record_retry_outcome(future, retry_futures, retry_results, still_failed)
-        return retry_results, still_failed  # Return results and failures.
-
-    @staticmethod
-    def _submit_synthetic_retries(executor, failed_devices, connection_semaphore):
-        """Submit retry calls for every failed device and return the future->device map."""
-        return {  # Map futures to devices.
-            executor.submit(
-                GatewayTestExporter.fetch_synthetic_test_stats_with_retry,
-                device_info,
-                max_retries=FAST_MODE_RETRY_MAX_RETRIES,
-                connection_semaphore=connection_semaphore,
-            ): device_info
-            for device_info in failed_devices
-        }
-
-    @staticmethod
-    def _record_retry_outcome(future, retry_futures, retry_results, still_failed):
-        """Inspect one future's result, append to the matching bucket, log the outcome."""
-        device_info = retry_futures[future]  # Resolve the device info.
-        try:
-            result = future.result()  # Read the result.
-            if result:  # Have a result.
-                retry_results.append(result)  # Collect it.
-                logging.info(" FAST RETRY OK: %s", device_info[2])  # Log retry success.
-            else:
-                still_failed.append(device_info)  # Still failed.
-                logging.error(" FAST RETRY FAIL: %s", device_info[2])  # Log retry failure.
-        except Exception as exception:  # Retry raised.
-            still_failed.append(device_info)  # Still failed.
-            logging.error(" FAST RETRY EXC: %s -> %s", device_info[2], exception)  # Log exception.
-
-    @staticmethod
-    def _run_synthetic_sequential_path(gateway_devices, all_stats):
-        """Sequential processing with adaptive rate limiting (original behavior)."""
-        smoothed = None  # No smoothed delay yet.
-        for device_info in tqdm(  # type: ignore[no-untyped-call]
-            gateway_devices, desc="Gateway Devices", unit="device"
-        ):
-            result = GatewayTestExporter.fetch_synthetic_test_stats_with_retry(
-                device_info, max_retries=FastModeSequentialMaxRetries.VALUE
-            )
-            if result:  # Have a result.
-                all_stats.append(result)  # Collect it.
-            smoothed, delay = RateLimitingUtils.get_rate_limited_delay(  # type: ignore[no-untyped-call]
-                smoothed, apisession, _api_usage_cache
-            )
-            logging.info("[INFO] Sleeping for %.2fs.", delay)  # Log the sleep.
-            time.sleep(delay)  # Pace the API.
-
-    @staticmethod
-    def _export_synthetic_results(all_stats, gateway_devices):
-        """Write the aggregated stats to CSV + log totals (or warn when empty)."""
-        if not all_stats:  # No results.
-            logging.warning(" No synthetic test results found. CSV not created.")  # Warn.
-            print("! No synthetic test results found. CSV not created.")  # Tell the user.
-            return  # Nothing to write.
-        filename = "AllGatewaySyntheticTests.csv"  # Build the CSV name.
-        flattened = DataProcessingUtils.flatten_nested_fields(all_stats)  # Flatten nested fields.
-        sanitized = DataProcessingUtils.escape_multiline(flattened)  # type: ignore[no-untyped-call]
-        DataExporter.write_with_format_selection(sanitized, filename)  # type: ignore[no-untyped-call]
-        print(f"! {len(all_stats)} gateway synthetic test results exported to {filename}")  # Tell user.
-        logging.info("! Synthetic test results saved to %s (%s records).", filename, len(all_stats))
-        logging.info(  # Log the optimization summary.
-            "! API Optimization: Saved %s listSiteDevices calls by using cached inventory",
-            len(gateway_devices),
-        )
-
-    @staticmethod
-    def test_results_by_site(fast: bool = False) -> None:  # Export tests by site.
-        """Delegator: all logic lives in src/refactors/serial_cc/test_results_by_site.py."""
-        from src.refactors.serial_cc.test_results_by_site import GatewayTestResultsService  # noqa: PLC0415
-
-        GatewayTestResultsService.execute(fast=fast)  # Delegate to extracted service; keeps CC at A(1)
+# GatewayTestExporter moved to src/export/gateway_test_exporter.py (1013 SC-001 position 37)  # noqa: E501
 
 
 class GatewayStatsExporter:  # Gateway stats delegators.
